restructure data constants
