+++ src/constants/data.constants.ts
@@ -1,0 +1,186 @@
+/**
+ * This file stores large lists of data
+ */
+
+// movie and Tv genres
+export const GENRE_MOVIE_TV = [
+  'Action',
+  'Adventure',
+  'Animation',
+  'Comedy',
+  'Crime',
+  'Documentary',
+  'Drama',
+  'Fantasy',
+  'Historical',
+  'Horror',
+  'Mystery',
+  'Romance',
+  'Sci-Fi',
+  'Thriller',
+  'War',
+  'Western',
+  'Superhero',
+  'Family',
+  'Musical',
+  'Biography',
+  'Sport',
+  'Noir',
+  'Reality',
+  'Talk Show',
+  'Game Show',
+  'News',
+  'Cooking',
+  'Kids',
+  'Indie',
+  'Strategy',
+  'Puzzle',
+  'Simulation',
+  'Fighting',
+  'Racing',
+  'Survival',
+  'Sandbox',
+  'Science Fiction',
+];
+
+//languages available
+export const LANGUAGES = [
+  'English',
+  'Hindi',
+  'Spanish',
+  'Mandarin',
+  'Arabic',
+  'Bengali',
+  'Portuguese',
+  'Russian',
+  'Japanese',
+  'Punjabi',
+  'German',
+  'Javanese',
+  'French',
+  'Telugu',
+  'Vietnamese',
+  'Marathi',
+  'Korean',
+  'Tamil',
+  'Turkish',
+  'Polish',
+  'Thai',
+  'Urdu',
+  'Greek',
+  'Malayalam',
+  'Romanian',
+  'Swedish',
+  'Hungarian',
+  'Italian',
+  'Dutch',
+  'Czech',
+  'Finnish',
+  'Norwegian',
+  'Danish',
+  'Bulgarian',
+  'Lithuanian',
+  'Hebrew',
+  'Croatian',
+  'Slovak',
+  'Slovenian',
+  'Estonian',
+  'Latvian',
+  'Maltese',
+];
+
+//movie , Tv show tags
+export const TAGS = [
+  'Based on true Story',
+  'Cult Classic',
+  'Binge-worthy',
+  'Award-winning',
+  'Coming of Age',
+  'Post-Apocalyptic',
+  'Dystopian',
+  'Feel-Good',
+  'Dark',
+  'Slow Burn',
+  'Fast-paced',
+  'Character-Driven',
+  'Plot Twist',
+  'Underrated',
+  'Critically Acclaimed',
+  'Soundtrack Heavy',
+  'Visual Masterpiece',
+  'Mind-Bending',
+  'Satirical',
+  'Non-linear Story',
+  'Multi-language',
+  'Psychological',
+  'Time Travel',
+  'Violent',
+  'Family Friendly',
+  'Adult',
+  'Romantic',
+  'Political',
+  'Fantasy World',
+  'Supernatural',
+  'Historical Fiction',
+  'Episodic',
+  'Serialized',
+  'Minimalist',
+  'Dialogue-heavy',
+  'Shounen',
+  'Sequel',
+  'Demon',
+  'Surveillance',
+  'Cybersecurity',
+  'Attack',
+  'Privacy',
+  'Government',
+  'Cop',
+  'Teacher',
+  'Trance',
+  'Witchcraft',
+  'Multiple storylines',
+  'Spell',
+  'Aggressive',
+  'Antagonistic',
+  'Exorcism',
+  'Haunted house',
+  'Religion',
+  'Wedding',
+  'Spirit',
+  'Ghost',
+  'Martial arts',
+  'Kung fu',
+  'Detective',
+  'Kidnapping',
+  'Heist',
+  'Cars',
+  'Orphan',
+  'Thief',
+  'Revenge',
+  'Dark comedy',
+  'Assassin',
+  'Reboot',
+  'Friendship',
+  'Politics',
+  'History',
+  'Suspense',
+  'Supervillain',
+  'Super power',
+  'Comeback',
+  'Exhilarated',
+  'Motorsport',
+  'Terrorist plot',
+  'Infiltration',
+  'Spy',
+  'Artificial intelligence',
+  'Female protagonist',
+  'K-pop',
+  'Fandom',
+  '3d animation',
+  'Save the world',
+  'Remake',
+  'Battle',
+  'Warrior',
+  'Cult',
+  'Gloomy',
+];
